push threadconfig into worker classes. Remove circular dependencies
